商品出品機能(修正2)
