+++ app/javascript/channels/index.js
@@ -4,14 +4,3 @@
 const channels = require.context('.', true, /_channel\.js$/)
 channels.keys().forEach(channels)
 
-window.addEventListener('load', () => {
-  const priceInput = document.getElementById("item-price");
-  priceInput.addEventListener("input", () => {
-    const addTaxDom = document.getElementById("add-tax-price");
-    // console.log(priceInput.innerHTML);
-    addTaxDom.innerHTML = Math.floor(priceInput.value * 0.1 );
-    const addPofitDom = document.getElementById("profit");
-    // console.log(priceInput.innerHTML);
-    addPofitDom.innerHTML = Math.floor(priceInput.value - Math.floor(priceInput.value * 0.1 ))
-  })
-});
